Version 3
Now works across PCs on the same Workgroup using SSH keys and an Admin account. Comparison of installed programs has yet to be updated for this new version.

CSV output is a bit funky with a blank line inbetween each entry, I'm aware of it and hoping it'll be an easy problem to fix.

Removed Testing instructions as they would be extremely lengthy and would require the setup of multiple VMs and SSH servers.